Refactor commonalities of RecoTrackSelector and RecoTrackRefSelector to RecoTrackSelectorBase
I will soon add a third variant, and without the refactoring (as
suggested by a comment in RecoTrackRefSelector) the code would become
even more messy.
